+++ Enumerable/DefaultIfEmpty/DefaultIfEmptyEnumerable.cs
@@ -532,7 +532,7 @@ namespace pcysl5edgo.Collections.LINQ
             where TPredicate : unmanaged, IRefFunc<TSource, bool>
             => this.Count(predicate);
 
-        public bool TryGetElementAt(int index, out TSource element)
+        public bool TryGetElementAt(long index, out TSource element)
             => this.TryGetElementAt<DefaultIfEmptyEnumerable<TEnumerable, TEnumerator, TSource>, Enumerator, TSource>(index, out element);
 
         public bool TryGetFirst(out TSource first)
